Fixed issue #1515. Blank spaces no longer appear as search results.
Fixed an issue where blank spaces would be shown as a search result. 

This would arise when there is no available matching_text field for the result. 
As of now, an absence of matching_text will default to a portion of the additional_text field being displayed if available. 
This functionality will not work for pluslets due to their additional_text field being HTML and not plaintext.