added toggle functionality
Had to create a new class "toggle-on" and "toggle-off" piggybacking off of the css style for "ctag-on" and "ctag-off".

The issue with using ctag as a class in this case, necessitating a new blanket class,  is lack of control over when to set it on or off due to the css hover event style associated with the ctag.