hide description type dialog if there is no description, snapshot update
When there is no description, the description radio buttons are now hidden from view and are only enabled when a description is provided.

The saveSnapshot command is reduced to two instances of usage in order to quantity the entire transformation of the widget after a dialog commit as one instance of change in the editor's undo/redo history.