fix issue where faq would error if question field is too long
When submitting an faq, an error would occur if the question for the faq exceeded 255 characters.
The culprit is ultimately the recordTitle field for chchchanges table which only allows 255 characters.

The solution is to shorten the string to fit the column.

UtilityService is created here to implement the cleanString function.