fix optimistic update bug where pluslet across different sections
Fixed a bug where dragging a pluslet across different sections would error the optimistic update portion because of read-only property errors with the pluslet objects.

Solution was to use immer's produce function on the Pluslet objects themselves and move the splicing of the reordered Pluslet within the produce function as well.